Improve namespaces.
