+++ src/Event/AfterLogin.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\User\CurrentUser\Event;
+namespace Yiisoft\User\Event;
 
 use Yiisoft\Auth\IdentityInterface;
 
-final class AfterLogout
+final class AfterLogin
 {
     private IdentityInterface $identity;
