+++ src/Event/BeforeLogout.php
@@ -2,17 +2,28 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\User\CurrentUser\Event;
+namespace Yiisoft\User\Event;
 
 use Yiisoft\Auth\IdentityInterface;
 
-final class AfterLogin
+final class BeforeLogout
 {
     private IdentityInterface $identity;
+    private bool $isValid = true;
 
     public function __construct(IdentityInterface $identity)
     {
         $this->identity = $identity;
+    }
+
+    public function invalidate(): void
+    {
+        $this->isValid = false;
+    }
+
+    public function isValid(): bool
+    {
+        return $this->isValid;
     }
 
     public function getIdentity(): IdentityInterface
